% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textyle.R
\name{textyle}
\alias{textyle}
\title{Animate an text element}
\usage{
textyle(
  element,
  color = "coral",
  easing = "easeInCubic",
  transition = 1,
  duration = 400,
  delay = 100,
  class = "ex2",
  ...
)
}
\arguments{
\item{element}{An element.}

\item{color}{string. Color of the callback}

\item{easing}{string. An effect from \href{http://gsgd.co.uk/sandbox/jquery/easing/}{jquery-easing}.}

\item{transition}{integrer. Transition in seconds}

\item{duration}{integrer. Duration}

\item{delay}{integrer. Delay}

\item{class}{string. Class of the effect}

\item{...}{Additional element}
}
\description{
Animate an text element
}
\examples{
library(htmltools)

textyle(h1("TITLE"))

}
