% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_textyle.R
\name{use_textyle}
\alias{use_textyle}
\title{Load Textyle dependencies}
\usage{
use_textyle()
}
\description{
Load textyle dependencies, i.e textylejs, textylecss and jquery.easing
}
\examples{
use_textyle()

}
